Cleaneup after rebase
