+++ models/demos/deepseek_v3_b1/weights/transforms/attention.py
@@ -21,6 +21,7 @@ from models.demos.deepseek_v3_b1.weights.specs.overlap_configs import (
     KVB12_PROJ_SINGLE_DEVICE_OVERLAP_SPEC,
     O_PROJ_GATE_MM_RMSNORM_GAMMA_SINGLE_DEVICE_OVERLAP_SPEC,
     QAB_KVA_PROJ_SINGLE_DEVICE_OVERLAP_SPEC,
+    QAB_KVA_PROJ_SingleDeviceOverlapSpec,
 )
 from models.demos.deepseek_v3_b1.weights.transforms.moe import _tp_factors
 
@@ -65,6 +66,55 @@ def preprocess_kv_b12(
     return {"kv_b1_proj": kv_b1, "kv_b2_proj": kv_b2_pre}
 
 
+def _make_q_ab_kv_a_overlap_entries(
+    cfg: QAB_KVA_PROJ_SingleDeviceOverlapSpec,
+    mesh_shape: tuple[int, int],
+    q_a_proj_weights: torch.Tensor,
+    q_b_proj_weights: torch.Tensor,
+    kv_a_proj_weights: torch.Tensor,
+    dtype: ttnn.DataType,
+) -> list[OverlapEntry]:
+    """Validate MLA proj weights and build three :class:`OverlapEntry` items (no I/O)."""
+    q_b_tp = cfg.q_b_shard_spec.tp(mesh_shape)
+
+    assert (
+        q_a_proj_weights.shape == cfg.q_a_proj_shape
+    ), f"q_a_proj_weights must be {cfg.q_a_proj_shape}, got {tuple(q_a_proj_weights.shape)}"
+    expected_q_b_shape = (cfg.q_b_proj_shape[0], cfg.q_b_proj_shape[1] * q_b_tp)
+    assert (
+        tuple(q_b_proj_weights.shape) == expected_q_b_shape
+    ), f"q_b_proj_weights must be {expected_q_b_shape}, got {tuple(q_b_proj_weights.shape)}"
+    assert (
+        kv_a_proj_weights.shape == cfg.kv_a_proj_shape
+    ), f"kv_a_proj_weights must be {cfg.kv_a_proj_shape}, got {tuple(kv_a_proj_weights.shape)}"
+
+    q_a_packed = cfg.shuffle_q_a(q_a_proj_weights)
+    kv_reordered = cfg.shuffle_kv_a(kv_a_proj_weights)
+
+    q_b_shuffled_slices = [
+        cfg.shuffle_q_b(cfg.get_q_b_slice(q_b_proj_weights, tp_idx, mesh_shape)) for tp_idx in range(q_b_tp)
+    ]
+    q_b_preprocessed = torch.cat(q_b_shuffled_slices, dim=1) if q_b_tp > 1 else q_b_shuffled_slices[0]
+
+    return [
+        OverlapEntry(
+            "q_a_proj",
+            q_a_packed,
+            replace(cfg.q_a_shard_spec, raw_tensor_shape=tuple(q_a_packed.shape), dtype=dtype),
+        ),
+        OverlapEntry(
+            "q_b_proj",
+            q_b_preprocessed,
+            replace(cfg.q_b_shard_spec, raw_tensor_shape=tuple(q_b_preprocessed.shape), dtype=dtype),
+        ),
+        OverlapEntry(
+            "kv_a_proj",
+            kv_reordered,
+            replace(cfg.kv_a_shard_spec, raw_tensor_shape=tuple(kv_reordered.shape), dtype=dtype),
+        ),
+    ]
+
+
 def fuse_q_ab_kv_a(
     q_a_proj_weights: torch.Tensor,
     q_b_proj_weights: torch.Tensor,
@@ -77,36 +127,10 @@ def fuse_q_ab_kv_a(
     """Fuse q_a, q_b, and kv_a projection weights into one overlapped buffer."""
     cfg = QAB_KVA_PROJ_SINGLE_DEVICE_OVERLAP_SPEC
     mesh_shape = (device.shape[0], device.shape[1]) if device.get_num_devices() > 1 else (1, 1)
-    preprocessed = preprocess_q_ab_kv_a(q_a_proj_weights, q_b_proj_weights, kv_a_proj_weights, mesh_shape)
-    q_a_packed = preprocessed["q_a_proj"]
-    q_b_pre = preprocessed["q_b_proj"]
-    kv_reordered = preprocessed["kv_a_proj"]
-
-    return overlap_tensors(
-        [
-            [
-                OverlapEntry(
-                    "q_a_proj",
-                    q_a_packed,
-                    replace(cfg.q_a_shard_spec, raw_tensor_shape=tuple(q_a_packed.shape), dtype=dtype),
-                ),
-                OverlapEntry(
-                    "q_b_proj",
-                    q_b_pre,
-                    replace(cfg.q_b_shard_spec, raw_tensor_shape=tuple(q_b_pre.shape), dtype=dtype),
-                ),
-            ],
-            [
-                OverlapEntry(
-                    "kv_a_proj",
-                    kv_reordered,
-                    replace(cfg.kv_a_shard_spec, raw_tensor_shape=tuple(kv_reordered.shape), dtype=dtype),
-                ),
-            ],
-        ],
-        device=device,
-        move_to_device=move_to_device,
+    entries = _make_q_ab_kv_a_overlap_entries(
+        cfg, mesh_shape, q_a_proj_weights, q_b_proj_weights, kv_a_proj_weights, dtype
     )
+    return overlap_tensors(entries, device=device, move_to_device=move_to_device)
 
 
 def fuse_o_proj_gate_mm_norms(
@@ -126,20 +150,84 @@ def fuse_o_proj_gate_mm_norms(
 
     return overlap_tensors(
         [
-            [
-                OverlapEntry(
-                    "o_proj",
-                    o_proj_weights,
-                    replace(cfg.o_proj, raw_tensor_shape=tuple(o_proj_weights.shape), dtype=o_proj_dtype),
-                )
-            ],
-            [OverlapEntry("gate_mm", gate_mm_weights, cfg.gate_mm)],
-            [
-                OverlapEntry("attn_norm", attn_norm, cfg.attn_norm),
-                OverlapEntry("q_norm", q_norm, cfg.q_norm),
-                OverlapEntry("ffn_norm", ffn_norm, cfg.ffn_norm),
-            ],
-            [OverlapEntry("kv_norm", kv_norm, cfg.kv_norm)],
+            OverlapEntry(
+                "o_proj",
+                o_proj_weights,
+                replace(cfg.o_proj, raw_tensor_shape=tuple(o_proj_weights.shape), dtype=o_proj_dtype),
+            ),
+            OverlapEntry("gate_mm", gate_mm_weights, cfg.gate_mm),
+            OverlapEntry("attn_norm", attn_norm, cfg.attn_norm),
+            OverlapEntry("q_norm", q_norm, cfg.q_norm),
+            OverlapEntry("ffn_norm", ffn_norm, cfg.ffn_norm),
+            OverlapEntry("kv_norm", kv_norm, cfg.kv_norm),
+        ],
+        device=device,
+        move_to_device=move_to_device,
+    )
+
+
+def fuse_o_proj_tp4_shuffled_gate_mm_norms_q_ab_kv_a(
+    o_proj_weights: torch.Tensor,
+    gate_mm_weights: torch.Tensor,
+    attn_norm: torch.Tensor,
+    q_norm: torch.Tensor,
+    kv_norm: torch.Tensor,
+    ffn_norm: torch.Tensor,
+    q_a_proj_weights: torch.Tensor,
+    q_b_proj_weights: torch.Tensor,
+    kv_a_proj_weights: torch.Tensor,
+    device,
+    *,
+    o_proj_dtype: ttnn.DataType = ttnn.bfloat8_b,
+    mla_proj_dtype: ttnn.DataType = ttnn.bfloat8_b,
+    move_to_device: bool = True,
+) -> dict[str, OverlappedTensor]:
+    """Fuse TP4 ``shuffle_q_a`` o_proj, gate_mm, norms, and q_a / q_b / kv_a into one L1 buffer.
+
+    Requires a **4×2** mesh.  ``o_proj_weights`` shape ``(16384, 7168)``.
+    """
+    o_cfg = O_PROJ_GATE_MM_RMSNORM_GAMMA_SINGLE_DEVICE_OVERLAP_SPEC
+    q_cfg = QAB_KVA_PROJ_SINGLE_DEVICE_OVERLAP_SPEC
+    mesh_shape = (device.shape[0], device.shape[1])
+    if mesh_shape != (4, 2):
+        raise ValueError(
+            "fuse_o_proj_tp4_shuffled_gate_mm_norms_q_ab_kv_a requires a 4x2 mesh; "
+            f"got {mesh_shape[0]}x{mesh_shape[1]}"
+        )
+    if tuple(o_proj_weights.shape) != (16384, 7168):
+        raise ValueError(
+            "o_proj_weights must have shape (16384, 7168) for TP4+shuffle layout; " f"got {tuple(o_proj_weights.shape)}"
+        )
+
+    device_grid = device.compute_with_storage_grid_size()
+    q_ab_bb = q_cfg.q_a_shard_spec.core_range_set.bounding_box()
+    kv_bb = q_cfg.kv_a_shard_spec.core_range_set.bounding_box()
+    required_rows = max(q_ab_bb.end.y, kv_bb.end.y) + 1
+    required_cols = max(q_ab_bb.end.x, kv_bb.end.x) + 1
+    assert device_grid.y >= required_rows, f"Device grid needs at least {required_rows} rows, got {device_grid.y}"
+    assert device_grid.x >= required_cols, f"Device grid needs at least {required_cols} cols, got {device_grid.x}"
+
+    o_packed = o_cfg.pack_o_proj_weights_tp4_shuffled(o_proj_weights)
+    o_spec = replace(
+        o_cfg.o_proj,
+        raw_tensor_shape=tuple(o_packed.shape),
+        dtype=o_proj_dtype,
+        tp_dim=(1, 0),
+    )
+
+    q_entries = _make_q_ab_kv_a_overlap_entries(
+        q_cfg, mesh_shape, q_a_proj_weights, q_b_proj_weights, kv_a_proj_weights, mla_proj_dtype
+    )
+
+    return overlap_tensors(
+        [
+            OverlapEntry("o_proj", o_packed, o_spec),
+            OverlapEntry("gate_mm", gate_mm_weights, o_cfg.gate_mm),
+            OverlapEntry("attn_norm", attn_norm, o_cfg.attn_norm),
+            OverlapEntry("q_norm", q_norm, o_cfg.q_norm),
+            OverlapEntry("ffn_norm", ffn_norm, o_cfg.ffn_norm),
+            OverlapEntry("kv_norm", kv_norm, o_cfg.kv_norm),
+            *q_entries,
         ],
         device=device,
         move_to_device=move_to_device,
@@ -163,25 +251,21 @@ def fuse_kv_b12(
 
     return overlap_tensors(
         [
-            [
-                OverlapEntry(
-                    "kv_b1_proj",
-                    kv_b1,
-                    replace(cfg.kv_b1_shard_spec, raw_tensor_shape=tuple(kv_b1.shape), dtype=dtype),
+            OverlapEntry(
+                "kv_b1_proj",
+                kv_b1,
+                replace(cfg.kv_b1_shard_spec, raw_tensor_shape=tuple(kv_b1.shape), dtype=dtype),
+            ),
+            OverlapEntry(
+                "kv_b2_proj",
+                kv_b2_pre,
+                replace(
+                    cfg.kv_b2_shard_spec,
+                    raw_tensor_shape=tuple(kv_b2_pre.shape),
+                    dtype=dtype,
+                    logical_tensor_shape=cfg.kv_b2_proj_shape,
                 ),
-            ],
-            [
-                OverlapEntry(
-                    "kv_b2_proj",
-                    kv_b2_pre,
-                    replace(
-                        cfg.kv_b2_shard_spec,
-                        raw_tensor_shape=tuple(kv_b2_pre.shape),
-                        dtype=dtype,
-                        logical_tensor_shape=cfg.kv_b2_proj_shape,
-                    ),
-                ),
-            ],
+            ),
         ],
         device=device,
         move_to_device=move_to_device,
